+++ packages/pypi/src/gira_cli/__init__.py
@@ -1,0 +1,3 @@
+from .installer import main
+
+__all__ = ["main"]